+++ mailit/views.py
@@ -3,10 +3,14 @@ from .models import MailItTemplate
 from .forms import MailitTemplateForm
 from subdomains.utils import reverse
 from django.shortcuts import get_object_or_404
-from django.http import Http404
+from django.http import Http404, HttpResponse
 from instance.models import WriteItInstance
 from django.contrib.auth.decorators import login_required
 from django.utils.decorators import method_decorator
+from django.views import View
+from mailit.exceptions import CouldNotFindIdentifier, TemporaryFailure
+from mailit.answer import OutboundMessageAnswer
+from mailit.bin.handleemail import EmailHandler
 
 
 class MailitTemplateUpdateView(UpdateView):
@@ -32,3 +36,17 @@ class MailitTemplateUpdateView(UpdateView):
 
     def get_success_url(self):
         return reverse('writeitinstance_template_update', subdomain=self.writeitinstance.slug)
+
+
+class IncomingMail(View):
+    def post(self, request):
+        print(request.body)  # TODO: clean up print statement
+        handler = EmailHandler(answer_class=OutboundMessageAnswer)
+        try:
+            answer = handler.handle(request.body)
+            answer.send_back()
+        except CouldNotFindIdentifier:
+            pass
+        except TemporaryFailure:
+            pass
+        return HttpResponse()
